Deduplicate Manual Document Service Registries
(Based on the changes made to the Manual Service Registries:
f4d70f400148f3d8bcc344eac236fbad9538a30c)

As described in f4d70f400148f3d8bcc344eac236fbad9538a30c, the big difference
between the previous incarnation of these service registries was the repository
each one used - with one service registry using a privileged repository and the
other using a restricted-by-organisation repository.

The two Manual Document Registries were previously offering a similar but not
identical set of services. These changes combine both sets of services in an
abstract registry and also put the config (of which repository the specialised
registry uses) in separate concrete classes, one for the privileged case and
one for the restricted case.
